move path_to_artifact to projectconfig
