+++ compiler/crates/relay-compiler/src/build_project/mod.rs
@@ -9,7 +9,6 @@
 //! watch mode or other state.
 
 mod artifact_generated_types;
-mod artifact_locator;
 pub mod artifact_writer;
 mod build_ir;
 pub mod build_schema;
@@ -30,7 +29,6 @@ use crate::errors::BuildProjectError;
 use crate::file_source::SourceControlUpdateStatus;
 use crate::{artifact_map::ArtifactMap, graphql_asts::GraphQLAsts};
 pub use artifact_generated_types::ArtifactGeneratedTypes;
-pub use artifact_locator::{create_path_for_artifact, path_for_artifact};
 use build_ir::BuildIRResult;
 pub use build_ir::SourceHashes;
 pub use build_schema::build_schema;
